refactor : 한번만 이동할 수 있는 기물들에 대하여 추상화하여 변경
